Sensor-simulator/pyproject.toml
[project]
name = "sensor-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bacpypes>=0.19.0",
    "paho-mqtt>=2.1.0",
    "pymodbus==3.6.6",
]

[tool.pytest.ini_options]
pythonpath = "."
